Remove unused rxkad_GetServerInfo() parameters
A few callers were giving pointers to rxkad_GetServerInfo() to extract
information the caller didn't actually need (that is, for the security
level and/or kvno for the conn). Change these callers to pass NULL for
the relevant parameters, to make it clearer in the code that this info
is not used.

Previous commit ae5dc207b9 (Remove unused expire times for
rxkad_GetServerInfo()) performed similar cleanup for expiration times
only.

Change-Id: I7b44a1f994fb2f297a3bfe857ac24b2359b708fe
Reviewed-on: https://gerrit.openafs.org/14027
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>