rxgk: Convert RXGK_ autoconf vars to IF_RXGK
Currently, we have a few autoconf vars that are set conditionally,
depending on whether or not rxgk is enabled (such as RXGK_LIBS,
RXGK_LIBS_RPC, etc). This works, but defining the values for these
vars is done very far away from where they are used, and this means
that for any part of a Makefile that is conditional on rxgk being
enabled, we must define a new autoconf var in src/cf/gss.m4.

To shift these definitions closer to where they are used, and arguably
simplify the process, convert these variables to use an
@IF_RXGK@-style approach, like we do for @IF_SHARED@ and other similar
conditional features. That is, instead of doing this:

    # gss.m4
    RXGK_LIBS=""
    AS_IF([...], [RXGK_LIBS="\$(top_builddir)/src/rxgk/liboafs_rxgk.la"])
    AC_SUBST([RXGK_LIBS])

    # src/foo/Makefile.in
    LT_deps = [...] @RXGK_LIBS@

We now have something like this:

    # gss.m4
    IF_RXGK="#"
    AS_IF([...], [IF_RXGK=""])
    AC_SUBST([IF_RXGK])

    # src/foo/Makefile.in
    @IF_RXGK@LT_deps_rxgk=$(top_builddir)/src/rxgk/liboafs_rxgk.la
    LT_deps = [...] $(LT_deps_rxgk)

Specifically, perform this conversion for: RXGK_LIBS, RXGK_LIBS_RPC,
RXGK_CFLAGS, and RXGK_GSSAPI_LIBS.

Provide a bit of documentation for this new subst var in the
OPENAFS_GSS autoconf function. Also document some of the other
rxgk-related autoconf bits while we're here to explain what all of the
vars mean.

Change-Id: I02db79473395d7e10281d5e5e9f8151121f999c8
Reviewed-on: https://gerrit.openafs.org/14029
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>